refactor(core): extract DU-stub walkers to du-stubs.ts (CORE-P3)
Move the `applyDuStubs` + `walkDuStubs` cluster out of
`create-form-store.ts` (~70 LOC) into a focused `du-stubs.ts` module.
The walkers are pure functions of `(schema, data, path, warned-set)`
— every dependency is a parameter, none capture closure state — so the
move is a verbatim relocation with no behavior delta.

Both call sites (`createFormStore` construction-time stubbing of the
authored defaults and the runtime `reshapeUnionAtPath`) keep their
existing semantics; `proto-pollution.test.ts` continues to exercise
the `__proto__` / `constructor` / `prototype` rejection through the
public hydration path, and `values-storage-shape.test.ts` +
`library-hardening-probes.test.ts` continue to cover the
disc-isn't-a-known-variant stubbing path. All 218 affected tests stay
green unchanged.

The host file drops the `isDangerousSegment` import (no longer used
locally) and picks up the `applyDuStubs` import from the new module;
the public surface is byte-identical.
